requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-cors-headers==3.7.0
Pillow==8.2.0
keras==2.4.3
pylint-django==2.4.4
tensorflow==2.5.0 